Rework UI, add search, build in assets (#12)
I got tired of X so I Y:

  X: the index page
  Y: so I made some bad bootstrap pages to make it prettier

  X: having to navigate to things by clicking
  Y: added a UI on top of the search API

  X: needing to manually copy assets into place for new installs
  Y: used [gobundle](http://github.com/falun/gobundle) to build in a standard set of assets

The MIME map stuff was added because I hit css/js problems w/
`X-Content-Type-Options=nosniff`. It was super lazily added/wired. When I
tested later with the MIME map removed it seemed to work though but I'm leaving
it in now because _shrug_ websites.